separate grphiql page for 3 components
