+++ src/components/Editor/Editor.scss
@@ -1,0 +1,41 @@
+.editor {
+  border: 1px solid rgb(17, 255, 0);
+  height: 100%;
+  width: 100%;
+  padding: 5px;
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+
+  .request {
+    width: 100%;
+    height: 100%;
+    border: 1px solid gray;
+    padding: 3px;
+    display: flex;
+    flex-direction: column;
+    gap: 3px;
+  }
+
+  .subquest {
+    border: 1px solid red;
+    width: 100%;
+    min-height: 100px;
+    padding: 3px;
+    display: flex;
+    flex-direction: column;
+    gap: 3px;
+
+    &_btnRow {
+      border: 1px solid rgb(0, 255, 234);
+      padding: 5px;
+      display: flex;
+      gap: 10px;
+    }
+    &_box {
+      border: 1px solid rgb(242, 0, 255);
+      height: 200px;
+      padding: 5px;
+    }
+  }
+}
