+++ src/components/Response/Response.scss
@@ -1,0 +1,6 @@
+.response {
+  border: 1px solid rgb(4, 0, 255);
+  height: 100%;
+  width: 100%;
+  padding: 5px;
+}
