+++ src/components/Response/Response.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import './Response.scss';
+
+function Response() {
+  return <section className="response">response</section>;
+}
+
+export default Response;
