+++ src/pages/Graphiql/Graphiql.scss
@@ -12,102 +12,4 @@
   @media (max-width: 1023px) {
     flex-direction: column;
   }
-
-  .docs {
-    border: 1px solid rgb(255, 0, 0);
-    height: 100%;
-    width: 33px;
-    padding: 3px;
-    padding-right: 30px;
-    position: relative;
-    transition: all 0.2s ease;
-    @media (max-width: 1023px) {
-      height: 70px;
-      width: 100%;
-    }
-    &._open {
-      width: 500px;
-      @media (max-width: 1023px) {
-        height: 200px;
-        width: 100%;
-      }
-    }
-
-    &_openBtn {
-      height: 30px;
-      width: 30px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      position: absolute;
-      right: 0px;
-      font-size: 20px;
-      font-weight: 700;
-    }
-
-    &_text {
-      border: 1px solid black;
-      height: 100%;
-      width: 100%;
-      display: flex;
-      transform: scale(0, 1);
-      padding: 3px;
-      overflow: auto;
-      flex-direction: column;
-      gap: 3px;
-      transition: all 0.2s ease;
-      &._open {
-        transform: scale(1, 1);
-      }
-    }
-  }
-
-  .editor {
-    border: 1px solid rgb(17, 255, 0);
-    height: 100%;
-    width: 100%;
-    padding: 5px;
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-
-    .request {
-      width: 100%;
-      height: 100%;
-      border: 1px solid gray;
-      padding: 3px;
-      display: flex;
-      flex-direction: column;
-      gap: 3px;
-    }
-
-    .subquest {
-      border: 1px solid red;
-      width: 100%;
-      min-height: 100px;
-      padding: 3px;
-      display: flex;
-      flex-direction: column;
-      gap: 3px;
-
-      &_btnRow {
-        border: 1px solid rgb(0, 255, 234);
-        padding: 5px;
-        display: flex;
-        gap: 10px;
-      }
-      &_box {
-        border: 1px solid rgb(242, 0, 255);
-        height: 200px;
-        padding: 5px;
-      }
-    }
-  }
-
-  .response {
-    border: 1px solid rgb(4, 0, 255);
-    height: 100%;
-    width: 100%;
-    padding: 5px;
-  }
 }
